code/WF2DH/pyproject.toml
[project]
name = "wf2dh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "configparser>=7.2.0",
    "fastapi[standard]>=0.128.0",
    "logging>=0.4.9.6",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
]
